# ENTREGABLES_SIGAI_SES/01_DOCUMENTACION_TECNICA/API_SPEC/openapi.yaml
openapi: 3.0.3
info:
  title: SIGAI-SES API
  version: "1.0.0"
  description: >
    API RESTful del Sistema Integral de Gestión de Activos e Inventario
    para Seguridad Electrónica Securitas Colombia.
  contact:
    name: Equipo de Desarrollo SIGAI-SES
    email: soporte@ses-securitas.com.co
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.ses.example.com/v1
    description: Producción
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido de /auth/login
  schemas:
...
    Cliente:
      type: object
      properties:
        id_cliente:
          type: integer
        nombre:
          type: string
        nit:
          type: string
        contacto:
          type: string
        email_contacto:
          type: string
        telefono:
          type: string
        ciudad:
          type: string
        tipo_cliente:
          type: string
          enum: [CORPORATIVO, INTERNO, GENERAL]
        created_at:
          type: string
          format: date-time

    # ==================== PROYECTOS ====================
    Proyecto:
      type: object
      properties:
        id_proyecto:
          type: integer
        id_cliente:
          type: integer
        nombre_proyecto:
          type: string
        centro_costos:
          type: string
        ubicacion:
          type: string
        estado:
          type: string
          enum: [ACTIVO, FINALIZADO, PAUSADO]
        fecha_inicio:
          type: string
          format: date
        fecha_fin_estimada:
          type: string
          format: date

    # ==================== PROVEEDORES ====================
...
    Garantia:
      type: object
      properties:
        id_garantia:
          type: integer
        id_activo:
          type: integer
        id_proveedor:
          type: integer
        numero_caso_interno:
          type: string
        rma_proveedor:
          type: string
        estado_proceso:
          type: string
          enum: [REGISTRADO, ENVIADO_PROVEEDOR, RECIBIDO_PROVEEDOR, RESUELTO_REEMPLAZADO, ENTREGADO_CLIENTE]
        falla_reportada:
          type: string
        tipo_resolucion:
          type: string
          enum: [REPARADO, REEMPLAZADO, SIN_COBERTURA, PENDIENTE]

    # ==================== MOVIMIENTOS ====================
    Movimiento:
      type: object
      properties:
        id_movimiento:
          type: integer
        id_usuario:
          type: integer
        id_item:
          type: integer
        id_activo:
          type: integer
        tipo_movimiento:
          type: string
          enum: [ENTRADA_COMPRA, SALIDA_INSTALACION, TRASLADO, DEVOLUCION, BAJA_DANO, AJUSTE, INGRESO_DESMONTE]
        cantidad:
          type: number
        origen:
          type: string
        destino:
          type: string
        fecha_movimiento:
          type: string
          format: date-time

    # ==================== ACTAS ENTREGA ====================
    ActaEntrega:
      type: object
      properties:
        id_acta:
          type: integer
        numero_acta:
          type: string
        id_usuario_tecnico:
          type: integer
        id_usuario_representante:
          type: integer
        id_proyecto:
          type: integer
        tipo_acta:
          type: string
          enum: [ENTREGA_EPP, ENTREGA_HERRAMIENTA, DESPACHO_PROYECTO, DEVOLUCION, INGRESO_DESMONTE]
        estado_acta:
          type: string
          enum: [BORRADOR, FIRMADA, ANULADA]
        fecha_entrega:
          type: string
          format: date-time
        url_pdf:
          type: string

    # ==================== ALERTAS ====================
...
security:
  - BearerAuth: []
paths:
...
  /business/clientes:
    get:
      summary: Listar clientes
      tags: [Business]
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      summary: Crear cliente
      tags: [Business]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                nit:
                  type: string
                contacto:
                  type: string
                email_contacto:
                  type: string
                telefono:
                  type: string
                ciudad:
                  type: string
                tipo_cliente:
                  type: string
                  default: CORPORATIVO
      responses:
        "201":
          description: Cliente creado
  /business/clientes/{id_cliente}:
    put:
      summary: Actualizar cliente
      tags: [Business]
      parameters:
        - name: id_cliente
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente actualizado
    delete:
      summary: Eliminar cliente
      tags: [Business]
      parameters:
        - name: id_cliente
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cliente eliminado
  /business/proyectos:
    get:
      summary: Listar proyectos
      tags: [Business]
      responses:
        "200":
          description: Lista de proyectos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proyecto"
    post:
      summary: Crear proyecto
      tags: [Business]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_cliente:
                  type: integer
                nombre_proyecto:
                  type: string
                centro_costos:
                  type: string
                ubicacion:
                  type: string
                estado:
                  type: string
                  default: ACTIVO
                fecha_inicio:
                  type: string
                  format: date
                fecha_fin_estimada:
                  type: string
                  format: date
      responses:
        "201":
          description: Proyecto creado
  /business/proyectos/{id_proyecto}:
    put:
      summary: Actualizar proyecto
      tags: [Business]
      parameters:
        - name: id_proyecto
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proyecto actualizado
    delete:
      summary: Eliminar proyecto
      tags: [Business]
      parameters:
        - name: id_proyecto
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Proyecto eliminado
...
  /business/garantias:
    get:
      summary: Listar garantias
      tags: [Business]
      responses:
        "200":
          description: Lista de garantias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Garantia"
    post:
      summary: Crear caso de garantia
      tags: [Business]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_activo:
                  type: integer
                id_proveedor:
                  type: integer
                falla_reportada:
                  type: string
                numero_caso_interno:
                  type: string
      responses:
        "201":
          description: Garantia creada
...
  /business/movimientos:
    get:
      summary: Listar movimientos de inventario
      tags: [Business]
      responses:
        "200":
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movimiento"
    post:
      summary: Crear movimiento de inventario
      tags: [Business]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_usuario:
                  type: integer
                id_item:
                  type: integer
                id_activo:
                  type: integer
                tipo_movimiento:
                  type: string
                cantidad:
                  type: number
                origen:
                  type: string
                destino:
                  type: string
      responses:
        "201":
          description: Movimiento creado
  /business/actas:
    get:
      summary: Listar actas de entrega
      tags: [Business]
      responses:
        "200":
          description: Lista de actas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActaEntrega"
    post:
      summary: Crear acta de entrega
      tags: [Business]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_usuario_tecnico:
                  type: integer
                id_usuario_representante:
                  type: integer
                id_proyecto:
                  type: integer
                tipo_acta:
                  type: string
                observaciones:
                  type: string
                detalles:
                  type: array
                  items:
                    type: object
                    properties:
                      id_item:
                        type: integer
                      id_activo:
                        type: integer
                      cantidad:
                        type: number
                      notas_estado:
                        type: string
      responses:
        "201":
          description: Acta creada
...